[jnc_ct] fix: LlvmIrBuilder::createSwitch takes const int64_t*; adjust call-sites accordingly

+++ src/jnc_ct/jnc_ct_ControlFlowMgr/jnc_ct_ControlFlowMgr_Eh.cpp
@@ -502,7 +502,7 @@ ControlFlowMgr::finalizeDisposableScope(Scope* scope) {
 	BasicBlock* switchBlock = m_currentBlock;
 
 	char buffer1[256];
-	sl::Array<intptr_t> levelArray(rc::BufKind_Stack, buffer1, sizeof(buffer1));
+	sl::Array<int64_t> levelArray(rc::BufKind_Stack, buffer1, sizeof(buffer1));
 	levelArray.setCount(count);
 
 	char buffer2[256];
